fix(migrations): cast int to int64 for gitlab SDK v2 ListOptions
The gitlab SDK v2 (gitlab.com/gitlab-org/api/client-go/v2) changed
ListOptions.Page and PerPage from int to int64. Add explicit casts to
fix compilation after merging main.

+++ services/migrations/gitlab.go
@@ -787,8 +787,8 @@ func (g *GitlabDownloader) GetOrgRepositories(ctx context.Context, groupName str
 
 	opt := &gitlab.ListGroupProjectsOptions{
 		ListOptions: gitlab.ListOptions{
-			Page:    page,
-			PerPage: perPage,
+			Page:    int64(page),
+			PerPage: int64(perPage),
 		},
 		IncludeSubGroups: new(bool),
 	}
